Initial configuration abstraction
